requirements.txt
audioread==2.1.8
backports.csv==1.0.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.23
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-haystack==1.8.5
elasticsearch==2.4.1
et-xmlfile==1.0.1
eventlet==0.25.1
fft==0.1
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
h5py==2.10.0
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
joblib==0.13.2
kiwisolver==1.1.0
librosa==0.7.0
llvmlite==0.29.0
matplotlib==3.0.3
monotonic==1.5
mutagen==1.42.0
numba==0.45.1
numpy==1.17.0
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.25.0
patsy==0.5.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.4
requests==2.22.0
resampy==0.2.1
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
SoundFile==0.10.2
sqlparse==0.3.0
statsmodels==0.10.1
tablib==0.13.0
urllib3==1.25.3
uWSGI==2.0.18
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
